Using radians in structs / ctypes. Fixing legacy conversion. Adding legacy bot for testing.
Fixing integration test.

+++ src/main/cpp/RLBotInterface/RLBotMessages/PacketStructs/LiveDataPacket.hpp
@@ -27,9 +27,9 @@ namespace PyStruct
 
 	struct Rotator
 	{
-		int Pitch;
-		int Yaw;
-		int Roll;
+		float Pitch;
+		float Yaw;
+		float Roll;
 	};
 
 	struct Color
